Change MeasurementUnit min max values fields types to float

+++ measurements/models.py
@@ -17,9 +17,9 @@ class MeasurementUnit(models.Model):
     uuid = models.UUIDField(primary_key=True, default=uuid4)
     name = models.CharField(max_length=settings.DEFAULT_CHAR_FIELD_LENGTH, unique=True)
     multiplier = models.IntegerField()
-    min_value = models.BigIntegerField()
+    min_value = models.FloatField()
     min_is_included = models.BooleanField(default=True)
-    max_value = models.BigIntegerField()
+    max_value = models.FloatField()
     max_is_included = models.BooleanField(default=True)
     group = models.ForeignKey(
         MeasurementGroup, on_delete=models.CASCADE, related_name='units'
